requirements.txt
#--- FastAPI and Server ---
fastapi
uvicorn[standard]
gunicorn
python-dotenv
python-multipart
argparse
requests
redis
passlib[bcrypt]

#--- Web UI ---
streamlit

#--- AI and Language Model ---
langchain
langchain-openai
langchain-community
langchain-chroma

#--- Vector Store ---
chromadb

#--- Document Processing ---
python-docx
pypdf
docx2txt
aiofiles
pymupdf
unstructured[local-inference]
pdfminer.six
openai>=1.3.0

#--- WhatsApp & Session Management ---
redis